Makefile:
-changed the date to be variable and moved it downwards into the archive
section.

New file:

Added a style file into the sub documents folder. This file will be filled with
commands predefined to change the header of the document. It serves also as
a space for the user to define commands and environments and change package
options. This leaves the class file in a clean state.