Changes on the header file:
Added a title, authors, affiliation and keywords to the header file. I made
also a changes to the table of the header from c to p{size} columns so the text
contents get wrapped properly. Additionally I changed the sizes of the title (to
LARGE), authors (normalsize), affiliation and keywords (footnotesize).