Layout changes:
Reduces the space between the header lines and the header content to 0.01 of
baselineskip. Changes the tabular columns from c to p{width} so the header
text content gets wrapped properly. The width changes on the minipages are
also related to that. Font size changes on the title (to LARGE), authors (to
normalsize), affiliation and keywords (to footnotesize).