Structural changes:
Removes the dist folder and contents again. Folders with example content and
temp files are added to the subdocuments folder. Two new methods are created
in the makefile where the first of which allows to start with an empty paper
altough it comes filled with example content by standard. The second one reverts
the to the delivery status with example contents.