Structural changes:
Adds a new folder to the graphics path to keep the dynamically created graphics
from the R chunks. The change in the global Knitr options takes care about that
and puts all of them there. This makes it easier to clean the document with the
makefile, which from now on also cleans that dynamic folder. The commit also
moves the example content to the new path and updates the PDF file.